+++ src/services/SubmissionConfirmationService.ts
@@ -57,7 +57,7 @@ export class SubmissionConfirmationService implements OnInit {
     }
 
 
-    public async generateConfirmationEntry(email: string, round: number): Promise<PendingEntryConfirmationModel> {
+    public async generateConfirmationEntry(round: number): Promise<PendingEntryConfirmationModel> {
         const confirmationModelRepository = this.ds.getRepository(PendingEntryConfirmationModel);
         const newEntry = this.ds.manager.create(PendingEntryConfirmationModel, {
             submissionId: round
